refactor: put meal related components into components/tracking folder
