+++ app/components/tracking/MacroSummary.tsx
@@ -2,8 +2,8 @@
 import React from "react";
 import { View, Text } from "react-native";
 import { Feather } from "@expo/vector-icons";
-import { useTheme } from "../context/ThemeContext";
-import { MacroData } from "../models/user";
+import { useTheme } from "../../context/ThemeContext";
+import { MacroData } from "../../models/user";
 import Svg, { Circle, G } from "react-native-svg";
 
 interface MacroSummaryProps {
