+++ app/components/tracking/MealList.tsx
@@ -2,9 +2,9 @@
 import React from "react";
 import { View, Text, Pressable } from "react-native";
 import { Feather } from "@expo/vector-icons";
-import { useTheme } from "../context/ThemeContext";
+import { useTheme } from "../../context/ThemeContext";
 import { useRouter } from "expo-router";
-import { FoodPortion } from "../models/food";
+import { FoodPortion } from "../../models/food";
 
 interface MealListProps {
   meals: {
